Add files via upload
This script will connect to the Conductor API, save the running configuration with the date, show peers and their status, as well as show peer path statistics for every peer.  Useful for pre and post network checks.   It will prompt for base url to the Conductor, username and password using getpass.